ETWeb/requirements.txt
six
python==3.7.2
django==3.0.8
channels==2.4.0
channels_redis
djangorestframework==3.11.0
psycopg2
Pillow